Error handling
New error: There was a bug in findStations(), in that one line was copied twice over (L146) and in another line (L151), the variable (correctedLCDVar_noRepeats[qq]) was not indexed, not an error for 1 variable but would be an issue if there were more variables. Known error: must sort dates so that the function will work.